# Improved fast serializer code generation logic for Enum type