--- a/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/Array_of_record_GenericSerializer_1629046702287533603.java
+++ b/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/Array_of_record_GenericSerializer_1629046702287533603.java
@@ -3,10 +3,7 @@ package com.linkedin.avro.fastserde.generated.serialization.AVRO_1_4;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import com.linkedin.avro.fastserde.FastSerializer;
-import org.apache.avro.Schema;
 import org.apache.avro.generic.IndexedRecord;
 import org.apache.avro.io.Encoder;
 import org.apache.avro.util.Utf8;
@@ -15,7 +12,6 @@ public class Array_of_record_GenericSerializer_1629046702287533603
     implements FastSerializer<List<IndexedRecord>>
 {
 
-    private Map<Long, Schema> enumSchemaMap = new ConcurrentHashMap<Long, Schema>();
 
     public void serialize(List<IndexedRecord> data, Encoder encoder)
         throws IOException