--- a/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerializer_6137374763587019173.java
+++ b/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerializer_6137374763587019173.java
@@ -3,8 +3,6 @@ package com.linkedin.avro.fastserde.generated.serialization.AVRO_1_4;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import com.linkedin.avro.fastserde.FastSerializer;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.IndexedRecord;
@@ -14,7 +12,7 @@ public class FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerialize
     implements FastSerializer<IndexedRecord>
 {
 
-    private Map<Long, Schema> enumSchemaMap = new ConcurrentHashMap<Long, Schema>();
+    private final Schema testEnumEnumSchema69 = Schema.parse("{\"type\":\"enum\",\"name\":\"testEnum\",\"namespace\":\"com.adpilot.utils.generated.avro\",\"symbols\":[\"A\",\"B\"]}");
 
     public void serialize(IndexedRecord data, Encoder encoder)
         throws IOException
@@ -26,11 +24,6 @@ public class FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerialize
     public void serializeFastGenericSerializerGeneratorTest_shouldWriteEnum68(IndexedRecord data, Encoder encoder)
         throws IOException
     {
-        Schema testEnumEnumSchema69 = enumSchemaMap.get(-3346156575824446940L);
-        if (null == testEnumEnumSchema69) {
-            testEnumEnumSchema69 = Schema.parse("{\"type\":\"enum\",\"name\":\"testEnum\",\"namespace\":\"com.adpilot.utils.generated.avro\",\"symbols\":[\"A\",\"B\"]}");
-            enumSchemaMap.put(-3346156575824446940L, testEnumEnumSchema69);
-        }
         (encoder).writeEnum(testEnumEnumSchema69 .getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) data.get(0)).toString()));
         org.apache.avro.generic.GenericData.EnumSymbol testEnumUnion70 = ((org.apache.avro.generic.GenericData.EnumSymbol) data.get(1));
         if (testEnumUnion70 == null) {
@@ -39,7 +32,7 @@ public class FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerialize
         } else {
             if (testEnumUnion70 instanceof org.apache.avro.generic.GenericData.EnumSymbol) {
                 (encoder).writeIndex(1);
-                (encoder).writeEnum(enumSchemaMap.get(-3346156575824446940L).getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) testEnumUnion70).toString()));
+                (encoder).writeEnum(testEnumEnumSchema69 .getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) testEnumUnion70).toString()));
             }
         }
         List<org.apache.avro.generic.GenericData.EnumSymbol> testEnumArray71 = ((List<org.apache.avro.generic.GenericData.EnumSymbol> ) data.get(2));
@@ -50,7 +43,7 @@ public class FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerialize
             (encoder).setItemCount(testEnumArray71 .size());
             for (int counter72 = 0; (counter72 <((List<org.apache.avro.generic.GenericData.EnumSymbol> ) testEnumArray71).size()); counter72 ++) {
                 (encoder).startItem();
-                (encoder).writeEnum(enumSchemaMap.get(-3346156575824446940L).getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) testEnumArray71 .get(counter72)).toString()));
+                (encoder).writeEnum(testEnumEnumSchema69 .getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) testEnumArray71 .get(counter72)).toString()));
             }
         }
         (encoder).writeArrayEnd();
@@ -70,7 +63,7 @@ public class FastGenericSerializerGeneratorTest_shouldWriteEnum_GenericSerialize
                 } else {
                     if (union75 instanceof org.apache.avro.generic.GenericData.EnumSymbol) {
                         (encoder).writeIndex(1);
-                        (encoder).writeEnum(enumSchemaMap.get(-3346156575824446940L).getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) union75).toString()));
+                        (encoder).writeEnum(testEnumEnumSchema69 .getEnumOrdinal(((org.apache.avro.generic.GenericData.EnumSymbol) union75).toString()));
                     }
                 }
             }