--- a/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/FastGenericSerializerGeneratorTest_shouldWriteFixed_GenericSerializer_8889056593487745201.java
+++ b/avro-fastserde/src/codegen/java/com/linkedin/avro/fastserde/generated/serialization/AVRO_1_4/FastGenericSerializerGeneratorTest_shouldWriteFixed_GenericSerializer_8889056593487745201.java
@@ -3,10 +3,7 @@ package com.linkedin.avro.fastserde.generated.serialization.AVRO_1_4;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import com.linkedin.avro.fastserde.FastSerializer;
-import org.apache.avro.Schema;
 import org.apache.avro.generic.IndexedRecord;
 import org.apache.avro.io.Encoder;
 
@@ -14,7 +11,6 @@ public class FastGenericSerializerGeneratorTest_shouldWriteFixed_GenericSerializ
     implements FastSerializer<IndexedRecord>
 {
 
-    private Map<Long, Schema> enumSchemaMap = new ConcurrentHashMap<Long, Schema>();
 
     public void serialize(IndexedRecord data, Encoder encoder)
         throws IOException