Improved fast serialization speed by avoiding the cost of storing and retrieving Enum schemas
in a Hashmap in Avro 1.4. We have seen ~44% improvement of serialization latency for Enum array.

This PR helps to resolve issue #50 

JMH benchmark results of fast serialization time of an 200 elements EnumArray under Avro 1.4

**Before** 
```
FastAvroSerdesBenchmark.testFastAvroSerialization                                   avgt   10  12054.915 ± 1852.177   ns/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.alloc.rate                    avgt   10    255.063 ±   42.062  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.alloc.rate.norm               avgt   10   3352.001 ±    0.001    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Eden_Space           avgt   10    272.156 ±  362.710  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Eden_Space.norm      avgt   10   3489.396 ± 4814.784    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Survivor_Space       avgt   10      0.103 ±    0.477  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Survivor_Space.norm  avgt   10      1.505 ±    6.959    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.count                         avgt   10      6.000             counts
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.time                          avgt   10     46.000                 ms
```

**After**
```
FastAvroSerdesBenchmark.testFastAvroSerialization                                   avgt   10  6709.067 ±  225.951   ns/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.alloc.rate                    avgt   10   453.950 ±   14.888  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.alloc.rate.norm               avgt   10  3352.000 ±    0.001    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Eden_Space           avgt   10   452.412 ±  247.752  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Eden_Space.norm      avgt   10  3337.223 ± 1836.797    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Survivor_Space       avgt   10     0.104 ±    0.480  MB/sec
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.churn.PS_Survivor_Space.norm  avgt   10     0.802 ±    3.719    B/op
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.count                         avgt   10     9.000             counts
FastAvroSerdesBenchmark.testFastAvroSerialization:·gc.time                          avgt   10    37.000                 ms
```