@FelixGV 2020-05-29
Cool result! And how does it compare with vanilla Avro now? Is it equal or better?

Can we get a diff of the generated code? Actually, I'm wondering... should we make the build trigger the code-gen for a set of schemas we care about, and check it under source control (not in the main resource, but perhaps in tests or in some kind of `audit` resource). If we're going to make it part of the workflow to review the generated code, we might as well think of a way to really institutionalize it properly...