@volauvent 2020-05-29
Fast-avro 1.4 serialization speed is ~10% faster than vanilla 1.4 now for Enum array. Below is the benchmark result of vanilla avro:
```
FastAvroSerdesBenchmark.testAvroSerialization                                       avgt   10  7432.819 ±  279.983   ns/op
FastAvroSerdesBenchmark.testAvroSerialization:·gc.alloc.rate                        avgt   10   410.783 ±   15.542  MB/sec
FastAvroSerdesBenchmark.testAvroSerialization:·gc.alloc.rate.norm                   avgt   10  3360.000 ±    0.001    B/op
FastAvroSerdesBenchmark.testAvroSerialization:·gc.churn.PS_Eden_Space               avgt   10   402.187 ±  326.169  MB/sec
FastAvroSerdesBenchmark.testAvroSerialization:·gc.churn.PS_Eden_Space.norm          avgt   10  3303.211 ± 2694.281    B/op
FastAvroSerdesBenchmark.testAvroSerialization:·gc.churn.PS_Survivor_Space           avgt   10     0.024 ±    0.094  MB/sec
FastAvroSerdesBenchmark.testAvroSerialization:·gc.churn.PS_Survivor_Space.norm      avgt   10     0.201 ±    0.791    B/op
FastAvroSerdesBenchmark.testAvroSerialization:·gc.count                             avgt   10     8.000             counts
FastAvroSerdesBenchmark.testAvroSerialization:·gc.time                              avgt   10    22.000                 ms
```

Here are fast-serializers of EnumArray schema:
- new fast-serializer before this PR:
https://gist.github.com/volauvent/d1fd5d25788fd8bdd33ba5bcf0ddfd27

- old fast-serializer after this PR:
https://gist.github.com/volauvent/78a117c5ffe0f76ef8ace15f632157f3

Like the idea of source control of code-gen to review related changes more carefully and properly.