@FelixGV 2020-06-01
Thanks for the gist... the change looks good. It looks quite weird to have a map lookup/set in this code. I wonder why it was done like this in the first place, and if that's a hint that there may be edge cases where this is important (I can't think of any, though).

Also, you'll notice that there is yet another map lookup hidden in the `getEnumOrdinal` function. This is kind of silly. If the `EnumSymbol` class kept an ordinal field within it, the lookup would be unnecessary. This is non-trivial to change because it would require extending or replacing some of the core Avro classes within `GenericData`, which we haven't done so far, but I had other reasons for wanting to do that in PR #45 so maybe we'll go there at some point... Anyhow, this is clearly out of scope from your current change.