@volauvent 2020-06-01
> I wonder why it was done like this in the first place, and if that's a hint that there may be edge cases where this is important (I can't think of any, though)

Support Avro 1.4 Enum type serialization here was initially introduced by @gaojieliu . Do you have any concern?